service_subsampling/requirement.txt
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-storage-blob==12.24.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
isodate==0.7.2
laspy==2.5.4
numpy==2.2.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
python-multipart==0.0.20
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
python-dotenv